TaxonomyScheme (#1)
This PR implements:
1. A simple class for a taxon (`Taxon`).
2. A family of classes for taxonomy (descendants of the abstract
`TaxonomyScheme`), primarily `PhylogeneticTaxonomyScheme`.
3. A family of classes for nomenclature (descendants of the abstract
`Nomenclature`), primarily `PangoSc2Nomenclature`.

Additional utilities are implemented where needed.

## Class structure
A line is drawn between nomenclature (which deals with names) and
taxonomy (which deals with relationships), which somewhat mirrors this
distinction in practice.

An individual `Taxon` is naive, only a `TaxonomyScheme` can tell us how
two `Taxon` objects are related.

### Nomenclature
The `Nomenclature` classes work with strings and serve to provide
information about the names of taxa.

The main class for current use is `PangoSc2Nomenclature`, and its main
use-cases are currently to:
1. Check the validity of the name of a taxon which nominally belongs to
the Pango SARS-CoV-2 nomenclature.
2. Obtain a phylogeny providing the taxonomic relationships between a
group of Pango SARS-CoV-2 taxa.

The class hierarchy here is:
1. `Nomenclature`, abstract, subclasses are assumed to be able to
determine whether a name is valid and some further information it
conveys, such as whether it indicates a hybrid taxon.
3. `AlgorithmicNomenclature` (mostly) abstract, assumes that names
encode taxonomic/evolutionary history and provides method for building a
phylogeny for a set of histories of those names.
4. `PangoLikeNomenclature`, (partially) abstract, provides many methods
for working with Pango nomenclature or similar schemes such as that
proposed for RSV (which is very similar to Pango but uses cumulative
aliases).
5. `PangoNomenclature`, not abstract (but not particularly intended for
end use either), asserts specific forms of more general objects (e.g.
that levels in a taxon's name are separated by ".", that names starting
with "X" indicate recombinants, and such).
6. `PangoSc2Nomenclature`, not abstract, sets member variables specific
to the SARS-CoV-2 Pango nomenclature (e.g. that there are 3 sublevels,
where the mpox Pango nomenclature uses 2)

### Taxonomy
The `Taxon` and `TaxonomyScheme` classes work with `Taxon` objects.

A `Taxon` is the basic unit to be used for taxonomy, primarily serving
as a way to link a name (e.g. "JN.1") to optional data (not currently
used in any functionality) and with an indicator of whether this taxon
is a "tip" or not. Where "tip" means, broadly, the most finely-available
unit of taxonomy available, something which could be observed in a
sample of lineages in a population.

The main current use for `TaxonomyScheme` classes is the
`PhylogeneticTaxonomyScheme`, which provides general methods for working
with taxonomy using a phylogenetic tree.

The class hierarchy in `TaxonomyScheme` is:
1. `TaxonomyScheme`, (mostly) abstract, subclasses are assumed to be
able to fulfill the general notions that a `Taxon` may have parents
(allowing multiple parents permits hybrid ancestry in principle) and
children, potentially multiple levels' worth for both of these, and that
one should be able to check if a taxon is valid and to check if a taxon
is the root (the taxon which contains all others).
2. `TreelikeTaxonomyScheme`, (mostly) abstract, specializes to the case
where all taxa have only one parent (and thus taxonomy is a tree), where
taxonomic notions are best-defined.
3. `PhylogeneticTaxonomyScheme`, not abstract, uses a user-provided
phylogeny to do taxonomy. The phylogeny must have labels for the taxon
represented by every node.

## Tests
Tests are based on `PangoSc2Nomenclature`, `Taxon`, and
`PhylogeneticTaxonomyScheme`. Tests are provided to check that the
taxonomic tree for Pango lineages can be built, that the Pango and
PangoLike name-wrangling functions work, and that the
`PhylogeneticTaxonomyScheme` works as anticipated.