Historical aggregation (#34)
## Background
Variant nowcasting and short-term forecasting predict the abundance of
different taxa. However, as evolution continues, new taxon names are
created, splitting up older ones.

Approximating what a nowcast/forecast would have looked like on some day
requires that we only model the taxa that were known on this as-of date.
Separately, but relatedly, evaluating/scoring _any_ nowcast/forecast
requires that all observations be labeled under the same set of
allowable taxa (the ones on the as-of date).

The tool [CladeTime](https://github.com/reichlab/cladetime) exists to do
this leveraging NextStrain's workflow. It does The Right Thing by
actually running the sequence assignment, enabling full [sequence-level
portability](https://github.com/CDCgov/cladecombiner/issues/30#issuecomment-2494040997)
by bridging the `sequence -> tip taxon` link.

However, when the requisite information has not been archived,
approximations are needed. Under certain assumptions (including
stability of the tree used for assigning labels), we can _approximate_
this with cladecombiner.

## This PR

This PR concerns this approximation, which we term "historical
aggregation." It enables using a tree of currently-recognized taxa, and
information about when taxa became recognized, to aggregate an input
taxon to its most-recent ancestor in the set of known taxa on a prior
date.

This resolves #9 in spirit and #27 completely.

To attempt to keep some potentially messy splits clean, I have:
- Assigned control/knowledge over versioning to `Nomenclature` rather
than `TaxonomyScheme` or some combination thereof. As the primary
utility lies in knowing what clade/lineage _names_ were assignable on
some date, I think this is sensible.
- Introduced a branch to the `Nomenclature` hierarchy
`HistoryAwareNomenclature` which evolving nomenclatures like Pango can
inherit from.

For the moment, a `HistoryAwareNomenclature` is just a thing that can
provide a tool (via `.get_versioner()`) which tells you if some name was
known on your as-of date. That notion is formalized with the
`NomenclatureVersioner` class which is a `Callable[[str], bool]` that
tells you such. For the moment, the only way to do this is with
`BruteForceNomenclatureVersioner` which uses the entire list of
recognized names on the as-of date to do this checking.

Historical aggregation is achieved by the `HistoricalAggregator` which
traverses a `PhylogeneticTaxonomyScheme` and uses the
`NomenclatureVersioner` to figure out what the tip-most taxa were in
that version. It then calls the `BasicPhylogeneticAggregator` to
actually do the aggregation.

Functionality for accessing and parsing the requisite files for Pango
lineages has been added, and all handling of Pango file reading has been
passed over to `PyGithub` for consistency as that is what we used for
the commit history traversing.

## Out of scope related issues

The aggregator workflow here highlights that we probably ought to
refactor what is currently called the `BasicPhylogeneticAggregator`.
That currently does the work of a `SerialAggregator` _and_ an
`Aggregator`.

We start to be confronted with the question of how explicit we want to
be about times and how mutable they should be. For now, I have attempted
to dodge this by adding the minimal amount required to make historical
aggregation possible, assuming that any instantiated taxonomy is current
(or at least, more current than the historical target). Eventually we
will want to consider whether we need to make Pango classes more aware
of when all their information refers to, so it's in-sync. And we may
need to consider whether the time of such a class is immutable (probably
should be) and whether we might want a means to spawn older
nomenclatures from newer ones
(`[SubclassOf]Nomenclature.reversion_as_of(date) -> Nomenclature`).

Now that we have access to all accepted Pango SARS-CoV-2 lineage names,
do we want to enable more strict validity checking? And if so, should
that be explicitly versioned?