Tools for working with Nextstrain SC2 clades (#33)
This PR serves to add tooling for working with Nextstrain SARS-CoV-2
clades.

The end result is an object provided to the user
`cladecombiner.nextstrain_sc2_nomenclature` that works pretty comparably
to `cladecombiner.pango_sc2_nomenclature`, in that it can do basic
`Nomenclature` things (nothing algorithmic) and it can get a taxonomy
tree for a set of taxa usable for aggregation.

Usage examples are provided in "Other nomenclatures and taxonomies" in
the documentation.

## `NextstrainLikeNomenclature`

This class is used to make the pre-baked
`cladecombiner.nextstrain_sc2_nomenclature` object for users.

It assumes that there is a GitHub repo with a file that defines the
master (taxonomy) tree, and another file that defines the master list of
recognized taxa. Functions to parse these are passed in at
initialization, leading to a relatively sparse class, and some pre-baked
versions for handling the Nextstrain SC2 clades.

To avoid immediately pinging GitHub for information upon importing
cladecombiner, most of the instance methods are wrapped in one of two
new decorators `ensure_taxa_known` or `ensure_tree_known`. These
intercept method calls that require the master taxon list or master
tree, and if that hasn't been accessed, grabs them. To keep these in
sync, the class has a single as-of date that passes to all PyGithub
calls to read files.

## Refactoring `Nomenclature`

I also added an `ArbitraryNomenclature` as the head of the
non-algorithmic branch of `Nomenclature`, from which
`NextstrainLikeNomenclature` descends.

I added a new `Nomenclature` subclass, `TreeProvider` which is a
`Nomenclature` that has a `.taxonomy_tree()` method like the Pango
branch of `Nomenclature`. This allows us to promise that other branches
of the `Nomenclature` tree can also provide a taxonomy tree for given
taxa.

## New (general) tree parsing/wrangling

The Nextstrain tree comes as a matrix specifying the parent taxon for
each taxon. Functionality has been added to `cladecombiner.tree_utils()`
to read this in and turn it into a tree. Tests thereof are added in this
PR.

Getting the taxonomy tree for a `NextstrainLikeNomenclature` is a
different problem than for an `AlgorithmicNomenclature` in that it
doesn't require building a tree but pruning one. Functionality for this
has also been added to `cladecombiner.tree_utils()` and tests thereof
are part of this PR.